fix: 홈 페이지 퍼블리싱 수정
