+++ lib/screen/home/home.dart
@@ -2,16 +2,14 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 
 import 'package:poorlex/widget/nav_bar.dart';
-import 'package:poorlex/widget/main/carousel_slider.dart';
-import 'package:poorlex/widget/main/main_bottom.dart';
-import 'package:poorlex/widget/main/show_modal.dart';
+import 'package:poorlex/widget/home/carousel_slider.dart';
+import 'package:poorlex/widget/home/home_bottom_button.dart';
+import 'package:poorlex/widget/home/show_modal.dart';
 
 import 'package:poorlex/controller/modal.dart';
 
 class Main extends StatefulWidget {
-  const Main({
-    super.key,
-  });
+  const Main({super.key});
 
   @override
   State<Main> createState() => _MainState();
@@ -51,10 +49,31 @@ class _MainState extends State<Main> {
 
   @override
   Widget build(BuildContext context) {
+    final double deviceWidth = MediaQuery.of(context).size.width;
+    final double deviceHeight = MediaQuery.of(context).size.height;
+
+    // 디바이스 높이에 따라 top, bottom 값을 설정합니다.
+    double topPosition;
+    double bottomPosition;
+
+    if (deviceHeight > 800) {
+      topPosition = 160.0;
+      bottomPosition = 100.0;
+    } else if (deviceHeight > 700) {
+      topPosition = 130.0;
+      bottomPosition = 75.0;
+    } else {
+      topPosition = 80.0;
+      bottomPosition = 50.0;
+    }
+
     return Scaffold(
       // 네비게이션 바
       appBar: NavBar(),
       body: Container(
+        width: MediaQuery.of(context).size.width,
+        height: MediaQuery.of(context).size.height,
+
         // 전체 배경
         decoration: BoxDecoration(
           image: DecorationImage(
@@ -64,39 +83,35 @@ class _MainState extends State<Main> {
         ),
 
         // 캐릭터 이미지, 배틀 리스트, 지출 추가하기 버튼
-        child: Container(
-          height: double.infinity,
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              Column(
+        child: Stack(
+          children: [
+            // 캐릭터 이미지
+            Positioned(
+              top: topPosition,
+              width: deviceWidth,
+              child: Row(
+                mainAxisAlignment: MainAxisAlignment.center,
                 children: [
-                  SizedBox(height: MediaQuery.of(context).size.height * 0.2),
-
-                  Column(
-                    children: [
-                      Container(
-                        width: 80,
-                        height: 94,
-                        child: Image.asset('assets/main_page/avatar_1.png'),
-                      ),
-                    ],
-                  ),
-
-                  SizedBox(height: MediaQuery.of(context).size.height * 0.05),
-
-                  // 참여 중인 배틀
                   Container(
-                    child: MainCarouselSlider(),
+                    width: 80,
+                    height: 94,
+                    child: Image.asset('assets/main_page/avatar_1.png'),
                   ),
                 ],
               ),
-              // 지출 추가하기 버튼
-              const MainBottom(),
-            ],
-          ),
+            ),
+
+            // 참여 중인 배틀 캐러셀
+            Positioned(
+              bottom: bottomPosition,
+              width: deviceWidth,
+              child: MainCarouselSlider(),
+            ),
+          ],
         ),
       ),
+      // 지출 추가 버튼
+      bottomNavigationBar: const MainBottom(),
     );
   }
 }
